requirements.txt
setuptools
oslo.config
python-glanceclient>=0.12,<=0.19
python-keystoneclient==1.6.0
python-novaclient>=2.0,<=2.26
python-neutronclient>=2.0,<=2.6
